Updates the UI for displaying total views count
FYI. This strays from the latest Sketch comps from Product as it deals
with potentially null total counts. The existing comps would stray the
block / mute button. This implementation builds off the pieces we had in
place previously, which handles those potentially missing values as well
as if the user is featured, has sharing etc...

[Finishes: #130300593](https://www.pivotaltracker.com/story/show/130300593)
[Finishes: #140886803](https://www.pivotaltracker.com/story/show/140886803)