Updates meta attributes for the video element
This doesn't seem to entirely fix the issue in the simulator but does provide
some improvements. In iOS 10, [Webkit upgraded their policy on playing
videos](https://webkit.org/blog/6784/new-video-policies-for-ios/)
:point_left::+1::book:. This update just implements the changes they
suggest.

React however, doesn't add the `muted` attribute to the dom element
(though viewing the controls does show it muted). Not sure if the
missing attribute is related to the issues I'm seeing in the simulator
or not. There is a related [issue in
react](https://github.com/facebook/react/issues/6544).

[#141788047](https://www.pivotaltracker.com/story/show/141788047)