Remove lineHeight from the blacklist and cleanup
Really the only line heights we care about are within our `minPixel`
range of `4` and those just carry through with the raw value. So for
`lineHeight` this works great, but for `margin` or other lower settings
we should ensure we are sending the required unit. Meh. Added a
shorthand of `lh` for `lineHeight` values similar to `width` and
`height`.

[#138191285](https://www.pivotaltracker.com/story/show/138191285)