Fix images in comments for mobile
Seriously WTF? The image elements get the correct width and height set
but for some reason are coming back all stretched. Setting the height to
auto in CSS fixes this but WTF? The height value on the element should
have precedence. WTF? We can't do this for all `ImageAttachments` since
this will break the reserved spacing while the image is loading. This
breaks the placeholder, but only for images when they are in comments
and only while they are loading (it gets a small placeholder since there
is no content). Small use case. Still WTF?

[Fixes: #143082473](https://www.pivotaltracker.com/story/show/143082473)