Update test:watch and add a test:watch:all script
Typically we want test:watch to just run the unit test. When the
development server is running we don't have the dist from build so the
server tests just fall over. We'll clean this up when we do the move to
webpack 2. This is way more messy than it needs to be.