Guard on scroll bottom launching the signup modal
If a logged out user scrolled to the bottom before the stream result had
loaded content a blank modal would appear. Seems to only be a Safari
issue. The guard here should protect against it. Not 100% sure this
fixes everything, but a place to start.

Also added a secondary check to ensure we have an authPromo...

[Fixes: #143465275](https://www.pivotaltracker.com/story/show/143465275)