Use split instead of replace to assemble iso responses
Thanks to a quirk of `String.replace`, a post including `$'` would
render weirdly because the match would get duplicated back into the body
a second time. See
https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/replace#Specifying_a_string_as_a_parameter
for an explanation of the mechanics.

Eventually, would be nice for the pipeline to return `&apos;` instead of
`'` which would get around this with fewer machinations, but this does
the job for now.

[Finishes #144326413](https://www.pivotaltracker.com/story/show/144326413)