Bust the iso cache when the build hash changes
This incorporates the build hash from Webpack into the cache key for
Memcache for iso responses, which will result in the cache getting blown
every time we do a deploy.

While there's a minor performance hit associated with this, it should
hopefully help our sanity on deploy as it will ensure that whatever
markup users are getting is reflecting the latest deployed code

See https://webpack.github.io/docs/long-term-caching.html for details on
the stats JSON that Webpack emits.

[Finishes #144559667](https://www.pivotaltracker.com/story/show/144559667)